Refactor CLI and implement interactive demo
