+++ go.mod
@@ -1,3 +1,12 @@
 module runix
 
 go 1.23.8
+
+require (
+	github.com/briandowns/spinner v1.22.0 // indirect
+	github.com/fatih/color v1.7.0 // indirect
+	github.com/mattn/go-colorable v0.1.2 // indirect
+	github.com/mattn/go-isatty v0.0.8 // indirect
+	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
+	golang.org/x/term v0.1.0 // indirect
+)
